+++ src/ru/nsu/ccfit/zuev/osu/Utils.java
@@ -57,9 +57,9 @@ public class Utils {
         pos.y += (Config.getRES_HEIGHT() - toRes(Constants.MAP_ACTUAL_HEIGHT)) / 2f;
         pos.x += (Config.getRES_WIDTH() - toRes(Constants.MAP_ACTUAL_WIDTH)) / 2f;
         if (GameHelper.isHardrock()) {
-            pos.y -= Config.getRES_HEIGHT() / 2;
+            pos.y -= (float) Config.getRES_HEIGHT() / 2;
             pos.y *= -1;
-            pos.y += Config.getRES_HEIGHT() / 2;
+            pos.y += (float) Config.getRES_HEIGHT() / 2;
         }
 		/*if (pos.y < 18) {
 			pos.y = 18;
@@ -75,9 +75,9 @@ public class Utils {
         pos.y += (Config.getRES_HEIGHT() - toRes(Constants.MAP_ACTUAL_HEIGHT)) / 2f;
         pos.x += (Config.getRES_WIDTH() - toRes(Constants.MAP_ACTUAL_WIDTH)) / 2f;
         if (GameHelper.isHardrock()) {
-            pos.y -= Config.getRES_HEIGHT() / 2;
+            pos.y -= (float) Config.getRES_HEIGHT() / 2;
             pos.y *= -1;
-            pos.y += Config.getRES_HEIGHT() / 2;
+            pos.y += (float) Config.getRES_HEIGHT() / 2;
         }
     }
